requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==5.0.4
django-appconf==1.0.6
django-environ==0.11.2
django-jazzmin==3.0.0
django-rest-swagger==2.2.0
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
executing==2.0.1
idna==3.7
inflection==0.5.1
ipdb==0.13.13
ipython==8.23.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.3
Markdown==3.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
openapi-codec==1.3.2
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==1.7.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
simplegeneric==0.8.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
